+++ ezdxf/dxfentity.py
@@ -6,7 +6,7 @@ from .lldxf.types import cast_tag_value, DXFTag
 from .lldxf.const import DXFStructureError, DXFInternalEzdxfError, DXFAttributeError, DXFInvalidLayerName
 from .lldxf.const import DXFKeyError, DXFValueError
 from .lldxf.validator import is_valid_layer_name
-from .lldxf.tags import Tags
+from .lldxf.tags import Tags, tuples2dxftags
 from .tools import set_flag_state
 
 ACAD_REACTORS = '{ACAD_REACTORS'
@@ -346,6 +346,7 @@ class DXFEntity(object):
         return self.tags.get_app_data_content(appid)
 
     def set_app_data(self, appid, app_data_tags):
+        app_data_tags = tuples2dxftags(app_data_tags)
         if self.tags.has_app_data(appid):
             appdata = self.tags.get_app_data(appid)
             appdata[1:-1] = app_data_tags
@@ -359,6 +360,7 @@ class DXFEntity(object):
         return Tags(self.tags.get_xdata(appid)[1:])  # without app id tag
 
     def set_xdata(self, appid, xdata_tags):
+        xdata_tags = tuples2dxftags(xdata_tags)
         if self.tags.has_xdata(appid):
             xdata = self.tags.get_xdata(appid)
             xdata[1:] = xdata_tags